code clean up, pod fixes, renamed Helper->Utils

+++ lib/Thruk/Controller/extinfo.pm
@@ -3,7 +3,6 @@ package Thruk::Controller::extinfo;
 use strict;
 use warnings;
 use parent 'Catalyst::Controller';
-use Thruk::Helper;
 
 =head1 NAME
 
@@ -70,6 +69,8 @@ sub index :Path :Args(0) :MyAction('AddDefaults') {
     $c->stash->{infoBoxTitle}   = $infoBoxTitle;
     $c->stash->{page}           = 'extinfo';
     $c->stash->{template}       = 'extinfo_type_'.$type.'.tt';
+
+    return 1;
 }
 
 
@@ -81,16 +82,18 @@ sub index :Path :Args(0) :MyAction('AddDefaults') {
 # create the downtimes page
 sub _process_comments_page {
     my ( $self, $c ) = @_;
-    $c->stash->{'hostcomments'}    = $c->{'live'}->selectall_arrayref("GET comments\n".Thruk::Helper::get_auth_filter($c, 'comments')."\nColumns: host_name id source type author comment entry_time entry_type expire_time expires\nFilter: service_description = ", { Slice => {} });
-    $c->stash->{'servicecomments'} = $c->{'live'}->selectall_arrayref("GET comments\n".Thruk::Helper::get_auth_filter($c, 'comments')."\nColumns: host_name service_description id source type author comment entry_time entry_type expire_time expires\nFilter: service_description != ", { Slice => {} });
+    $c->stash->{'hostcomments'}    = $c->{'live'}->selectall_arrayref("GET comments\n".Thruk::Utils::get_auth_filter($c, 'comments')."\nColumns: host_name id source type author comment entry_time entry_type expire_time expires\nFilter: service_description = ", { Slice => {} });
+    $c->stash->{'servicecomments'} = $c->{'live'}->selectall_arrayref("GET comments\n".Thruk::Utils::get_auth_filter($c, 'comments')."\nColumns: host_name service_description id source type author comment entry_time entry_type expire_time expires\nFilter: service_description != ", { Slice => {} });
+    return 1;
 }
 
 ##########################################################
 # create the downtimes page
 sub _process_downtimes_page {
     my ( $self, $c ) = @_;
-    $c->stash->{'hostdowntimes'}    = $c->{'live'}->selectall_arrayref("GET downtimes\n".Thruk::Helper::get_auth_filter($c, 'downtimes')."\nFilter: service_description = \nColumns: author comment end_time entry_time fixed host_name id start_time triggered_by", { Slice => {} });
-    $c->stash->{'servicedowntimes'} = $c->{'live'}->selectall_arrayref("GET downtimes\n".Thruk::Helper::get_auth_filter($c, 'downtimes')."\nFilter: service_description != \nColumns: author comment end_time entry_time fixed host_name id service_description start_time triggered_by", { Slice => {} });
+    $c->stash->{'hostdowntimes'}    = $c->{'live'}->selectall_arrayref("GET downtimes\n".Thruk::Utils::get_auth_filter($c, 'downtimes')."\nFilter: service_description = \nColumns: author comment end_time entry_time fixed host_name id start_time triggered_by", { Slice => {} });
+    $c->stash->{'servicedowntimes'} = $c->{'live'}->selectall_arrayref("GET downtimes\n".Thruk::Utils::get_auth_filter($c, 'downtimes')."\nFilter: service_description != \nColumns: author comment end_time entry_time fixed host_name id service_description start_time triggered_by", { Slice => {} });
+    return 1;
 }
 
 ##########################################################
@@ -101,8 +104,8 @@ sub _process_host_page {
 
     my $backend  = $c->{'request'}->{'parameters'}->{'backend'};
     my $hostname = $c->{'request'}->{'parameters'}->{'host'};
-    $c->detach('/error/index/5') unless defined $hostname;
-    my $hosts = $c->{'live'}->selectall_hashref("GET hosts\n".Thruk::Helper::get_auth_filter($c, 'hosts')."\nFilter: name = $hostname\nColumns: has_been_checked accept_passive_checks acknowledged action_url_expanded address alias checks_enabled check_type current_attempt current_notification_number event_handler_enabled execution_time flap_detection_enabled groups icon_image_expanded icon_image_alt is_executing is_flapping last_check last_notification last_state_change latency long_plugin_output max_check_attempts name next_check notes_expanded notes_url_expanded notifications_enabled obsess_over_host parents percent_state_change perf_data plugin_output scheduled_downtime_depth state state_type", 'peer_key', {AddPeer => 1});
+    return $c->detach('/error/index/5') unless defined $hostname;
+    my $hosts = $c->{'live'}->selectall_hashref("GET hosts\n".Thruk::Utils::get_auth_filter($c, 'hosts')."\nFilter: name = $hostname\nColumns: has_been_checked accept_passive_checks acknowledged action_url_expanded address alias checks_enabled check_type current_attempt current_notification_number event_handler_enabled execution_time flap_detection_enabled groups icon_image_expanded icon_image_alt is_executing is_flapping last_check last_notification last_state_change latency long_plugin_output max_check_attempts name next_check notes_expanded notes_url_expanded notifications_enabled obsess_over_host parents percent_state_change perf_data plugin_output scheduled_downtime_depth state state_type", 'peer_key', {AddPeer => 1});
 
     # we only got one host
     if(scalar keys %{$hosts} == 1) {
@@ -118,15 +121,17 @@ sub _process_host_page {
         }
     }
 
-    $c->detach('/error/index/5') unless defined $host;
+    return $c->detach('/error/index/5') unless defined $host;
 
     my @backends = keys %{$hosts};
     $self->_set_backend_selector($c, \@backends, $host->{'peer_key'});
 
     $c->stash->{'host'}     = $host;
-    my $comments            = $c->{'live'}->selectall_arrayref("GET comments\n".Thruk::Helper::get_auth_filter($c, 'comments')."\nFilter: host_name = $hostname\nFilter: service_description =\nColumns: author id comment entry_time entry_type expire_time expires persistent source", { Slice => 1 });
-    my $sortedcomments      = Thruk::Helper->sort($c, $comments, 'id', 'DESC');
+    my $comments            = $c->{'live'}->selectall_arrayref("GET comments\n".Thruk::Utils::get_auth_filter($c, 'comments')."\nFilter: host_name = $hostname\nFilter: service_description =\nColumns: author id comment entry_time entry_type expire_time expires persistent source", { Slice => 1 });
+    my $sortedcomments      = Thruk::Utils::sort($c, $comments, 'id', 'DESC');
     $c->stash->{'comments'} = $sortedcomments;
+
+    return 1;
 }
 
 ##########################################################
@@ -135,14 +140,15 @@ sub _process_hostgroup_cmd_page {
     my ( $self, $c ) = @_;
 
     my $hostgroup = $c->{'request'}->{'parameters'}->{'hostgroup'};
-    $c->detach('/error/index/5') unless defined $hostgroup;
+    return $c->detach('/error/index/5') unless defined $hostgroup;
 
-    my $groups = $c->{'live'}->selectall_hashref("GET hostgroups\n".Thruk::Helper::get_auth_filter($c, 'hostgroups')."\nColumns: name alias\nFilter: name = $hostgroup\nLimit: 1", 'name');
+    my $groups = $c->{'live'}->selectall_hashref("GET hostgroups\n".Thruk::Utils::get_auth_filter($c, 'hostgroups')."\nColumns: name alias\nFilter: name = $hostgroup\nLimit: 1", 'name');
     my @groups = values %{$groups};
-    $c->detach('/error/index/5') unless defined $groups[0];
+    return $c->detach('/error/index/5') unless defined $groups[0];
 
     $c->stash->{'hostgroup'}       = $groups[0]->{'name'};
     $c->stash->{'hostgroup_alias'} = $groups[0]->{'alias'};
+    return 1;
 }
 
 ##########################################################
@@ -153,12 +159,12 @@ sub _process_service_page {
     my $backend  = $c->{'request'}->{'parameters'}->{'backend'};
 
     my $hostname = $c->{'request'}->{'parameters'}->{'host'};
-    $c->detach('/error/index/5') unless defined $hostname;
+    return $c->detach('/error/index/5') unless defined $hostname;
 
     my $servicename = $c->{'request'}->{'parameters'}->{'service'};
-    $c->detach('/error/index/5') unless defined $servicename;
+    return $c->detach('/error/index/5') unless defined $servicename;
 
-    my $services = $c->{'live'}->selectall_hashref("GET services\n".Thruk::Helper::get_auth_filter($c, 'services')."\nFilter: host_name = $hostname\nFilter: description = $servicename\nColumns: has_been_checked accept_passive_checks acknowledged action_url_expanded checks_enabled check_type current_attempt current_notification_number description event_handler_enabled execution_time flap_detection_enabled groups host_address host_alias host_name icon_image_expanded icon_image_alt is_executing is_flapping last_check last_notification last_state_change latency long_plugin_output max_check_attempts next_check notes_expanded notes_url_expanded notifications_enabled obsess_over_service percent_state_change perf_data plugin_output scheduled_downtime_depth state state_type", 'peer_key', {AddPeer => 1});
+    my $services = $c->{'live'}->selectall_hashref("GET services\n".Thruk::Utils::get_auth_filter($c, 'services')."\nFilter: host_name = $hostname\nFilter: description = $servicename\nColumns: has_been_checked accept_passive_checks acknowledged action_url_expanded checks_enabled check_type current_attempt current_notification_number description event_handler_enabled execution_time flap_detection_enabled groups host_address host_alias host_name icon_image_expanded icon_image_alt is_executing is_flapping last_check last_notification last_state_change latency long_plugin_output max_check_attempts next_check notes_expanded notes_url_expanded notifications_enabled obsess_over_service percent_state_change perf_data plugin_output scheduled_downtime_depth state state_type", 'peer_key', {AddPeer => 1});
 
     # we only got one service
     if(scalar keys %{$services} == 1) {
@@ -174,15 +180,17 @@ sub _process_service_page {
         }
     }
 
-    $c->detach('/error/index/5') unless defined $service;
+    return $c->detach('/error/index/5') unless defined $service;
 
     my @backends = keys %{$services};
     $self->_set_backend_selector($c, \@backends, $service->{'peer_key'});
 
     $c->stash->{'service'}  = $service;
-    my $comments            = $c->{'live'}->selectall_arrayref("GET comments\n".Thruk::Helper::get_auth_filter($c, 'comments')."\nFilter: host_name = $hostname\nFilter: service_description = $servicename\nColumns: author id comment entry_time entry_type expire_time expires persistent source", { Slice => 1 });
-    my $sortedcomments      = Thruk::Helper->sort($c, $comments, 'id', 'DESC');
+    my $comments            = $c->{'live'}->selectall_arrayref("GET comments\n".Thruk::Utils::get_auth_filter($c, 'comments')."\nFilter: host_name = $hostname\nFilter: service_description = $servicename\nColumns: author id comment entry_time entry_type expire_time expires persistent source", { Slice => 1 });
+    my $sortedcomments      = Thruk::Utils::sort($c, $comments, 'id', 'DESC');
     $c->stash->{'comments'} = $sortedcomments;
+
+    return 1;
 }
 
 ##########################################################
@@ -191,14 +199,16 @@ sub _process_servicegroup_cmd_page {
     my ( $self, $c ) = @_;
 
     my $servicegroup = $c->{'request'}->{'parameters'}->{'servicegroup'};
-    $c->detach('/error/index/5') unless defined $servicegroup;
+    return $c->detach('/error/index/5') unless defined $servicegroup;
 
-    my $groups = $c->{'live'}->selectall_hashref("GET servicegroups\n".Thruk::Helper::get_auth_filter($c, 'servicegroups')."\nColumns: name alias\nFilter: name = $servicegroup\nLimit: 1", 'name');
+    my $groups = $c->{'live'}->selectall_hashref("GET servicegroups\n".Thruk::Utils::get_auth_filter($c, 'servicegroups')."\nColumns: name alias\nFilter: name = $servicegroup\nLimit: 1", 'name');
     my @groups = values %{$groups};
     $c->detach('/error/index/5') unless defined $groups[0];
 
     $c->stash->{'servicegroup'}       = $groups[0]->{'name'};
     $c->stash->{'servicegroup_alias'} = $groups[0]->{'alias'};
+
+    return 1;
 }
 
 ##########################################################
@@ -220,12 +230,14 @@ sub _process_scheduling_page {
     };
     $sortoption = 7 if !defined $sortoptions->{$sortoption};
 
-    my $services = $c->{'live'}->selectall_arrayref("GET services\n".Thruk::Helper::get_auth_filter($c, 'services')."\nColumns: host_name description next_check last_check check_options active_checks_enabled\nFilter: active_checks_enabled = 1\nFilter: check_options != 0\nOr: 2", { Slice => {} });
-    my $hosts    = $c->{'live'}->selectall_arrayref("GET hosts\n".Thruk::Helper::get_auth_filter($c, 'hosts')."\nColumns: name next_check last_check check_options active_checks_enabled\nFilter: active_checks_enabled = 1\nFilter: check_options != 0\nOr: 2", { Slice => {}, rename => { 'name' => 'host_name' } });
-    my $queue    = Thruk::Helper->sort($c, [@{$hosts}, @{$services}], $sortoptions->{$sortoption}->[0], $order);
+    my $services = $c->{'live'}->selectall_arrayref("GET services\n".Thruk::Utils::get_auth_filter($c, 'services')."\nColumns: host_name description next_check last_check check_options active_checks_enabled\nFilter: active_checks_enabled = 1\nFilter: check_options != 0\nOr: 2", { Slice => {} });
+    my $hosts    = $c->{'live'}->selectall_arrayref("GET hosts\n".Thruk::Utils::get_auth_filter($c, 'hosts')."\nColumns: name next_check last_check check_options active_checks_enabled\nFilter: active_checks_enabled = 1\nFilter: check_options != 0\nOr: 2", { Slice => {}, rename => { 'name' => 'host_name' } });
+    my $queue    = Thruk::Utils::sort($c, [@{$hosts}, @{$services}], $sortoptions->{$sortoption}->[0], $order);
     $c->stash->{'queue'}   = $queue;
     $c->stash->{'order'}   = $order;
     $c->stash->{'sortkey'} = $sortoptions->{$sortoption}->[1];
+
+    return 1;
 }
 
 
@@ -234,7 +246,8 @@ sub _process_scheduling_page {
 sub _process_process_info_page {
     my ( $self, $c ) = @_;
 
-    $c->detach('/error/index/1') unless $c->check_user_roles( "authorized_for_system_information" );
+    return $c->detach('/error/index/1') unless $c->check_user_roles( "authorized_for_system_information" );
+    return 1;
 }
 
 ##########################################################
@@ -242,11 +255,12 @@ sub _process_process_info_page {
 sub _process_perf_info_page {
     my ( $self, $c ) = @_;
 
-    my $stats      = Thruk::Helper->get_service_exectution_stats($c);
-    my $live_stats = $c->{'live'}->selectrow_arrayref("GET status\n".Thruk::Helper::get_auth_filter($c, 'status')."\nColumns: cached_log_messages connections connections_rate host_checks host_checks_rate requests requests_rate service_checks service_checks_rate neb_callbacks neb_callbacks_rate", { Slice => 1, Sum => 1 });
+    my $stats      = Thruk::Utils::get_service_exectution_stats($c);
+    my $live_stats = $c->{'live'}->selectrow_arrayref("GET status\n".Thruk::Utils::get_auth_filter($c, 'status')."\nColumns: cached_log_messages connections connections_rate host_checks host_checks_rate requests requests_rate service_checks service_checks_rate neb_callbacks neb_callbacks_rate", { Slice => 1, Sum => 1 });
 
     $c->stash->{'stats'}      = $stats;
     $c->stash->{'live_stats'} = $live_stats;
+    return 1;
 }
 
 ##########################################################
@@ -264,6 +278,7 @@ sub _set_backend_selector {
 
     $c->stash->{'backends'} = \@backends;
     $c->stash->{'backend'}  = $selected;
+    return 1;
 }
 
 =head1 AUTHOR
